+++ src/Linna/Cache/CacheFactory.php
@@ -11,62 +11,29 @@ declare(strict_types=1);
 
 namespace Linna\Cache;
 
+use Linna\Storage\AbstractStorageFactory;
 use Psr\SimpleCache\CacheInterface;
-use InvalidArgumentException;
 
 /**
  * Storage Factory.
  */
-class CacheFactory
+class CacheFactory extends AbstractStorageFactory
 {
-    /**
-     * @var string One of supported drivers
-     */
-    private $driver;
-
     /**
      * @var array Factory supported driver
      */
-    private $supportedDriver = [
+    protected $supportedDriver = [
         'disk'       => DiskCache::class,
         'memcached'  => MemcachedCache::class,
     ];
 
     /**
-     * @var array Options for the driver
-     */
-    private $options;
-
-    /**
-     * Constructor.
-     *
-     * @param string $driver
-     * @param array  $options
-     */
-    public function __construct(string $driver, array $options)
-    {
-        $this->driver = $driver;
-        $this->options = $options;
-    }
-
-    /**
      * Return Cache Resource.
-     *
-     * @throws InvalidArgumentException If required driver is not supported
      *
      * @return CacheInterface
      */
     public function get() : CacheInterface
     {
-        $driver = $this->driver;
-        $options = $this->options;
-
-        if (isset($this->supportedDriver[$driver])) {
-            $cache = $this->supportedDriver[$driver];
-
-            return new $cache($options);
-        }
-
-        throw new InvalidArgumentException("[$driver] not supported.");
+        return $this->returnStorageObject();
     }
 }